Added option to enable infoOnHover
Added a new option to enable the feature to sho inoboxes on hovering over
the objects. This is only customizable per map and not per object as I don't think its useful to have a map with different objects reacting in a  different way.

To turn on add infoonhover=on.

In code this is done by binding mouseover and mouseout only when this
value is true.